+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Intercept the [glibc] `64` variant functions `nftw64()`, `scandir64()` and
   `truncate64()`
 - Intercept the functions `acct()`, `catopen()`, `ftok()` and `mq_open()`
+- Intercept the Linux's functions `swapon()` and `swapoff()`
 
 ### Changed